[Rename] `command.ts` to `commandRegistrant.ts`
