+++ src/code/platform/command/common/commandRegistrant.ts
